Organize scheme-related components properly
