+++ src/lib/browse/colors.ts
@@ -1,6 +1,4 @@
-import { constructMatchExpression } from "lib/maplibre";
-import { type DataDrivenPropertyValueSpecification } from "maplibre-gl";
-import { fundingProgrammesForColouringAndFiltering } from "./data";
+import type { DataDrivenPropertyValueSpecification } from "maplibre-gl";
 
 // These are specific to the browse page
 export const colors = {
@@ -89,65 +87,3 @@ export const denseLineWidth: DataDrivenPropertyValueSpecification<number> = [
   18,
   20.0,
 ];
-
-export function styleByFundingProgramme(): [
-  DataDrivenPropertyValueSpecification<string>,
-  [string, string][],
-] {
-  let [colorMapping, legendRows] = getColorMappingAndLegend(
-    fundingProgrammesForColouringAndFiltering,
-    colors.funding_programmes,
-  );
-
-  let color = constructMatchExpression(
-    ["get", "funding_programme"],
-    colorMapping,
-    "grey",
-  );
-  return [color, legendRows];
-}
-
-export function styleByCurrentMilestone(): [
-  DataDrivenPropertyValueSpecification<string>,
-  [string, string][],
-] {
-  let stageGates = [
-    "removed",
-    "no data",
-    "not progressed",
-    "superseded",
-    "preliminary design completed",
-    "feasability design completed",
-    "detailed design completed",
-    "consruction started",
-    "construction completed",
-  ];
-
-  let [colorMapping, legendRows] = getColorMappingAndLegend(
-    stageGates,
-    colors.current_milestone,
-  );
-
-  let color = constructMatchExpression(
-    ["get", "current_milestone"],
-    colorMapping,
-    "grey",
-  );
-  return [color, legendRows];
-}
-
-function getColorMappingAndLegend(
-  keys: string[],
-  colorList: string[],
-): [{ [key: string]: string }, [string, string][]] {
-  let legendRows: [string, string][] = [];
-  let colorMapping: { [key: string]: string } = {};
-  let i = 0;
-  for (let x of keys) {
-    let color = colorList[i++ % colorList.length];
-    colorMapping[x] = color;
-    legendRows.push([x, color]);
-  }
-
-  return [colorMapping, legendRows];
-}
